refactor(results): Move QBField Type to Core Types for Reusability (#169)
